requirements.txt
flask
mysql-connector-python
python-dotenv
flask-bcrypt
pyjwt
